+++ lib/src/ui/has_widgets.dart
@@ -7,7 +7,7 @@ part of dart_web_toolkit_ui;
  * A widget that implements this interface contains [Widget] and can enumerate
  * them.
  */
-abstract class HasWidgets implements Diterable<Widget>{
+abstract class HasWidgets{
 
   /**
    * Adds a child widget.
@@ -30,6 +30,11 @@ abstract class HasWidgets implements Diterable<Widget>{
    * @return <code>true</code> if the widget was present
    */
   bool remove(Widget w);
+  
+  /**
+   * Returns an [Iterator] that iterates over this [Iterable] object.
+   */
+  Iterator<Widget> iterator();
 }
 
 /**