+++ lib/src/ui/widget_collection.dart
@@ -16,7 +16,7 @@ part of dart_web_toolkit_ui;
  * supports removal of widgets.
  * </p>
  */
-class WidgetCollection implements Diterable<Widget> {
+class WidgetCollection extends Iterable<Widget> {
 
   static int _INITIAL_SIZE = 4;
 
@@ -124,7 +124,7 @@ class WidgetCollection implements Diterable<Widget> {
    *
    * @return an iterator
    */
-  RemoveIterator<Widget> iterator() {
+  RemoveIterator<Widget> get iterator {
     return new WidgetIterator(this);
   }