+++ lib/util.dart
@@ -18,4 +18,3 @@ part 'src/util/uri_utils.dart';
 part 'src/util/safe_html.dart';
 part 'src/util/safe_html_utils.dart';
 part 'src/util/safe_html_string.dart';
-part 'src/util/diterable.dart';